refactor motor API: trying to test if C++ calls can work
turns out that we can get around the name mangling but
(1) adding the libraries we need to make C++ work blows up
the memory size to bigger than we can afford and (2) we're
still missing an implementantion of malloc() and free().

So...need to do a port of the API to straight C for efficiency!